+++ src/classes/http/NostoHttpRequestAdapterCurl.php
@@ -70,8 +70,7 @@ class NostoHttpRequestAdapterCurl extends NostoHttpRequestAdapter
                 CURLOPT_FRESH_CONNECT => 1,
                 CURLOPT_RETURNTRANSFER => 1,
                 CURLOPT_FORBID_REUSE => 1,
-                CURLOPT_TIMEOUT => 60,
-            )
+           )
         );
     }
 
@@ -83,12 +82,17 @@ class NostoHttpRequestAdapterCurl extends NostoHttpRequestAdapter
      */
     protected function send(array $curlOptions)
     {
-        $headers = $this->getHeaders();
-        if (!empty($headers)) {
-            $curlOptions[CURLOPT_HTTPHEADER] = $headers;
+        if (!empty($this->headers)) {
+            $curlOptions[CURLOPT_HTTPHEADER] = $this->headers;
         }
         if (!in_array(CURLOPT_USERAGENT, $curlOptions) && $this->userAgent) {
             $curlOptions[CURLOPT_USERAGENT] = $this->userAgent;
+        }
+        if (!in_array(CURLOPT_TIMEOUT, $curlOptions)) {
+            $curlOptions[CURLOPT_TIMEOUT] = NostoHttpRequest::$responseTimeout;
+        }
+        if (!in_array(CURLOPT_CONNECTTIMEOUT, $curlOptions)) {
+            $curlOptions[CURLOPT_CONNECTTIMEOUT] = NostoHttpRequest::$connectTimeout;
         }
         $ch = curl_init();
         curl_setopt_array($ch, $curlOptions);
@@ -116,7 +120,6 @@ class NostoHttpRequestAdapterCurl extends NostoHttpRequestAdapter
                 CURLOPT_FRESH_CONNECT => 1,
                 CURLOPT_RETURNTRANSFER => 1,
                 CURLOPT_FORBID_REUSE => 1,
-                CURLOPT_TIMEOUT => 60,
             )
         );
     }
@@ -136,7 +139,6 @@ class NostoHttpRequestAdapterCurl extends NostoHttpRequestAdapter
                 CURLOPT_FRESH_CONNECT => 1,
                 CURLOPT_RETURNTRANSFER => 1,
                 CURLOPT_FORBID_REUSE => 1,
-                CURLOPT_TIMEOUT => 60,
             )
         );
     }
@@ -155,7 +157,6 @@ class NostoHttpRequestAdapterCurl extends NostoHttpRequestAdapter
                 CURLOPT_FRESH_CONNECT => 1,
                 CURLOPT_RETURNTRANSFER => 1,
                 CURLOPT_FORBID_REUSE => 1,
-                CURLOPT_TIMEOUT => 60,
             )
         );
     }